feat: reorganize InitializeEscrowForm component and update import paths for better structure
